Former-commit-id: efb20d3b9de277543e3405a26f8e86acf2ecc812

+++ ide/src/org/alice/media/audio/FFmpegAudioConverter.java
@@ -111,17 +111,20 @@ public class FFmpegAudioConverter {
 		}
 	}
 
-	public static File convertAudioIfNecessary( AudioResource resource ) {
+	public static AudioResource convertAudioIfNecessary( AudioResource resource ) {
+		if( !AudioResourceConverter.needsConverting( resource, desiredFormat ) ) {
+			return resource;
+		}
 		try {
 			File file = createAudioFileToConvert( resource );
 			File outputFile;
 			outputFile = File.createTempFile( "outputFile", ".wav" );
 			String path = outputFile.getAbsolutePath();
 			outputFile.delete();
-			RuntimeUtilities.exec( ffmpegCommand, "-i", file.getAbsolutePath(), "-acodec", "pcm_s16le",
+			RuntimeUtilities.execSilent( ffmpegCommand, "-i", file.getAbsolutePath(), "-acodec", "pcm_s16le",
 					"-ar", String.valueOf( RATE_44 ), "-ac", "1", path );
 			File convertedFile = new File( path );
-			return convertedFile;
+			return new AudioResource( convertedFile );
 		} catch( IOException e ) {
 			e.printStackTrace();
 		}
